Extract Darwin "PlatformHelpers"
